it was the clock speed all along

+++ keyboards/rsheldiii/codename_busboy/matrix.c
@@ -65,7 +65,7 @@ void matrix_init_custom(void) {
     // TODO: initialize hardware here
 }
 
-static uint8_t scan_now = 0;
+static uint16_t scan_now = 0;
 
 // bool matrix_scan_custom(matrix_row_t current_matrix[]) {
 //     // dprintf("%lx\n", current_matrix[0]);
